Add .gitignore to ignore user-added modules
The idea is that users and developers can have additional modules installed in
the modules directory for testing without them accidentally getting
added/committed to the skeleton project. This makes it so that to add a new
module to the repository under modules/, you'll need to explicitly add it to the
modules/.gitignore.