Add jquery to layout by default
I'm not 100% sold on shipping this with the skeleton, as it's one more
thing people have to remove if they're not using it. However, it's
required to make the responsive layout work properly and having it
included makes ZendDeveloperTools work with the skeleton out of the box,
which is a plus.