Provide development mode
zend-development-mode has been a popular feature of Apigility, and a popular
standalone module. Unfortunately, its architecture uses the MVC <-> console
integration, which is problematic in the streamlined skeleton, which does not
include that out-of-the-box.

This patch creates development-mode as as standalone script in the project
`bin/` directory. Additionally, we add the `config/modules.config.php`, to push
modules into a separate file, and `config/development.config.php.dist`, to
provide a basic structure for segregated configuration. Finally, the bootstrap
is updated to test for the presence of `config/development.config.php` and merge
its contents with the application configuration.

Usage is:

```bash
$ ./bin/development-mode enable
```

Or, to disable:

```bash
$ ./bin/development-mode disable
```