fix: use script for updating .gitignore
Due to quoting issues between different operating systems, the script
for removing the `composer.lock` entry from the `.gitignore` file was
not working on Linux after a recent update to provide support for
Windows. To provide a truly cross-platform way to work, this patch
provides a script, `bin/update-gitignore`, which does the work. The
`composer.json` entry is updated to call that script, which deletes
itself on completion.